+++ Milles_Project1Library/Models/Shape.cs
@@ -14,9 +14,9 @@ namespace Milles_Project1Library.Models
 
         public string ShapeType { get; set; }
 
-        public double Area { get; set; }
+        public decimal Area { get; set; }
 
-        public double Perimeter { get; set; }
+        public decimal Perimeter { get; set; }
 
         public DateTime CalculationDate { get; set; }
     }
